backend/pyproject.toml
[project]
name = "laundry"
version = "0.1.0"
description = "Telegram Mini App + Bot for shared laundry room management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-telegram-bot>=21.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "qrcode[pil]>=7.4.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
